New analytics model, motbot reactions, and new advice load system
*New analytics model: Recent Cognitive Presence*
Detects the level of cognitive presence of a student on the whole platform. The Level of cognitive presence is divided in 4 distinct levels: 
 - Level 0: No Recent Accesses
 - Level 1: Recent Accesses
 - Level 2: Recent write activity
 - Level 3: Recent completions or rewarded certificates (TODO)

*Motbot reactions*
When the user performs an action that successfully completes an Intervention, like writing a forum post after the motbot intervened because of low social presence, then the motbot sends a short notification, containing a positive comment. In case of successful completions of no_recent_accesses or recent_cognitive_presence interventions, the motbot welcomes the user back and also includes a recommendation to continue with the course or activity the user last accessed.

*New advice: Last Stop*
This new advice will recommend to continue with the last activity the user has accessed. Only activities that have completion tracking enabled will be suggested. If the last accessed activity was already completed, the next uncompleted activity will be suggested instead. If there is no activity to suggest, the last accessed course will be suggested instead.

*New advice load system*
In a file advice.php inside the db directory of a plugin, developers now can associate advice with targets. On install or upgrading of the motbot plugin, these associations will be loaded and written to the database. That way other plugins can add new targets and advice, that the motbot can then use as well. These associations define what kind of advice the motbot will try to generate, depending on the kind of prediction/analytics target that triggered the intervention.